+++ core/src/main/java/react4j/core/React.java
@@ -1,10 +1,13 @@
 package react4j.core;
 
 import elemental2.core.JsArray;
+import java.util.List;
+import java.util.stream.Stream;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import jsinterop.annotations.JsOverlay;
 import jsinterop.annotations.JsPackage;
+import jsinterop.annotations.JsProperty;
 import jsinterop.annotations.JsType;
 import jsinterop.base.Js;
 import jsinterop.base.JsPropertyMap;
@@ -18,6 +21,13 @@ public final class React
   private React()
   {
   }
+
+  /**
+   * The magic component constructor function for Fragments.
+   */
+  @SuppressWarnings( "unused" )
+  @JsProperty( name = "Fragment" )
+  private static ComponentConstructorFunction Fragment;
 
   /**
    * Create a ReactElement for the specified React component with no props or children.
@@ -86,6 +96,57 @@ public final class React
   public static native ReactElement<ComponentConstructorFunction> createElement( @Nonnull ComponentConstructorFunction type,
                                                                                  @Nullable JsPropertyMap<Object> props,
                                                                                  @Nonnull ReactNode[] children );
+
+  /**
+   * Create a Fragment with the specified children.
+   *
+   * @param children the child nodes.
+   * @return a new React.Fragment object.
+   */
+  @JsOverlay
+  public static ReactNode createFragment( @Nonnull final ReactNode... children )
+  {
+    return createElement( Fragment, null, children );
+  }
+
+  /**
+   * Create a Fragment with the specified children.
+   *
+   * @param children the child nodes.
+   * @return a new React.Fragment object.
+   * @see #createElement(ComponentConstructorFunction, JsPropertyMap, ReactNode...)
+   */
+  @JsOverlay
+  public static ReactNode createFragment( @Nonnull final JsArray<ReactNode> children )
+  {
+    return createFragment( Js.<ReactNode[]>cast( children ) );
+  }
+
+  /**
+   * Create a Fragment with the specified children.
+   *
+   * @param children the child nodes.
+   * @return a new React.Fragment object.
+   * @see #createElement(ComponentConstructorFunction, JsPropertyMap, ReactNode...)
+   */
+  @JsOverlay
+  public static ReactNode createFragment( @Nonnull final List<? extends ReactNode> children )
+  {
+    return createFragment( children.stream() );
+  }
+
+  /**
+   * Create a Fragment with the specified children.
+   *
+   * @param children the child nodes.
+   * @return a new React.Fragment object.
+   * @see #createElement(ComponentConstructorFunction, JsPropertyMap, ReactNode...)
+   */
+  @JsOverlay
+  public static ReactNode createFragment( @Nonnull final Stream<? extends ReactNode> children )
+  {
+    return createFragment( children.toArray( ReactNode[]::new ) );
+  }
 
   /**
    * Clone and return a new ReactElement using element as the starting point. The resulting