+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ### Unreleased
 
+##### Added
+* **\[core\]** Add support for the `React.Fragment` node type introduced in React `v16.2.0` as described in
+  the [fragments]()https://reactjs.org/blog/2017/11/28/react-v16.2.0-fragment-support.html) section of the
+  ReactJS release nodes. This has been implemented by adding several overlay `React.createFragment(...)` methods.
+
 #### Changed
 * **\[core\]** Convert the `React.createElement(...,JsArray)` method from a native method to an overlay method
   to reduce size of generated code and add opportunities for optimization within the GWT compiler.